% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{char_scalar}
\alias{char_scalar}
\title{Convert a length-one character type in R to String and back}
\usage{
char_scalar(x)
}
\arguments{
\item{x}{a length-one character type in R}
}
\description{
Convert a length-one character type in R to String and back
}
