% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{char_str_vec}
\alias{char_str_vec}
\title{Convert a character vector in R to Vec<String> and back}
\usage{
char_str_vec()
}
\arguments{
\item{x}{a character vector}
}
\description{
Convert a character vector in R to Vec<String> and back
}
