% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dbls_named}
\alias{dbls_named}
\title{Adds an attribute to a vector of doubles}
\usage{
dbls_named(x)
}
\arguments{
\item{x}{Vector of doubles}
}
\description{
Adds an attribute to a vector of doubles
}
