% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{double_mat}
\alias{double_mat}
\title{Convert a numeric matrix RArray RArray<&\link{f64}, \link{usize;2}> and back}
\usage{
double_mat()
}
\arguments{
\item{x}{a numeric matrix}
}
\description{
Convert a numeric matrix RArray RArray<&\link{f64}, \link{usize;2}> and back
}
