% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Calculate Euclidean distance matrix
Test case adopted from \url{https://github.com/mikemahoney218/examplerust/blob/23d21b1ced4e24b7a7c00dd36290114dc1bbd113/src/rust/src/lib.rs#L5}}
\usage{
euclidean_dist(a)
}
\arguments{
\item{a}{: Matrix of real values or \code{NULL}}
}
\description{
Calculate Euclidean distance matrix
Test case adopted from \url{https://github.com/mikemahoney218/examplerust/blob/23d21b1ced4e24b7a7c00dd36290114dc1bbd113/src/rust/src/lib.rs#L5}
}
