% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{f__00__special_function_name}
\alias{f__00__special_function_name}
\alias{__00__special_function_name}
\title{Test wrapping of special function name}
\usage{
`__00__special_function_name`()
}
\description{
Test wrapping of special function name
}
