% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{false}
\alias{false}
\title{Combine raw identifiers (\code{r#}) as a function name and in arguments are parsed correctly.
See (Issue #582)\link{https://github.com/extendr/extendr/issues/528} for details.}
\usage{
false(type)
}
\arguments{
\item{type}{: i32 or \code{NULL}}
}
\description{
Combine raw identifiers (\code{r#}) as a function name and in arguments are parsed correctly.
See (Issue #582)\link{https://github.com/extendr/extendr/issues/528} for details.
}
