% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloextendr.R
\docType{package}
\name{helloextendr}
\alias{helloextendr}
\title{An example package demonstrating the use of the extendr crate from R.}
\description{
An example package demonstrating the use of the extendr crate from R.
}
