% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendrtests.R, R/wrappers.R
\name{int_scalar}
\alias{int_scalar}
\title{convert an int scalar to itself}
\usage{
int_scalar(x)

int_scalar(x)
}
\arguments{
\item{x}{a number}
}
\description{
convert an int scalar to itself

convert an int scalar to itself
}
