% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{list_string_hash}
\alias{list_string_hash}
\title{Convert a list to a HashMap<String, Robj> in rust and back. Does not preserve list order}
\usage{
list_string_hash()
}
\arguments{
\item{x}{a list}
}
\description{
Convert a list to a HashMap<String, Robj> in rust and back. Does not preserve list order
}
