% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-wrappers.R
\name{make_wrappers}
\alias{make_wrappers}
\title{Generate wrapper code}
\usage{
make_wrappers()
}
\description{
Run \code{make_wrappers()} to update the wrapper scripts for this package.
This autogenerates the file \code{R/extendr-wrappers.R}.
}
