% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{my_device}
\alias{my_device}
\title{Create a new device.}
\usage{
my_device(welcome_message)
}
\arguments{
\item{welcome_message}{A warm message to welcome you.}
}
\description{
Create a new device.
}
