% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{MyClass2}
\alias{MyClass2}
\alias{$.MyClass2}
\title{Class for testing (exported) (use_try_from)}
\format{
An object of class \code{environment} of length 6.
}
\usage{
MyClass2
}
\description{
Class for testing (exported) (use_try_from)
}
\examples{
x <- MyClass2$new()
x$a()
x$set_a(10)
x$a()
}
\keyword{datasets}
