% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{MyClassTryFrom}
\alias{MyClassTryFrom}
\alias{$.MyClassTryFrom}
\title{Class for testing (exported)}
\format{
An object of class \code{environment} of length 6.
}
\usage{
MyClassTryFrom
}
\description{
Class for testing (exported)
}
\examples{
x <- MyClass$new()
x$a()
x$set_a(10)
x$a()
}
\keyword{datasets}
