% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{MyClassUnexportedTryFrom}
\alias{MyClassUnexportedTryFrom}
\title{Class for testing (unexported)}
\format{
An object of class \code{environment} of length 2.
}
\usage{
MyClassUnexportedTryFrom
}
\description{
Class for testing (unexported)
}
\keyword{datasets}
