% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{MySubmoduleClass}
\alias{MySubmoduleClass}
\alias{$.MySubmoduleClass}
\title{Class for testing (exported)}
\format{
An object of class \code{environment} of length 3.
}
\usage{
MySubmoduleClass
}
\description{
Class for testing (exported)
}
\section{Methods}{

\subsection{Method \code{new}}{
Method for making a new object.
}

\subsection{Method \code{set_a}}{
Method for setting stuff.
\subsection{Arguments}{
\describe{
\item{\code{x}}{a number}
}}
}

\subsection{Method \code{a}}{
Method for getting stuff.
}
}

\examples{
x <- MySubmoduleClass$new()
x$a()
x$set_a(10)
x$a()
}
\keyword{datasets}
