% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{not_invisible_unit}
\alias{not_invisible_unit}
\title{Test function that is explicitly NOT invisible.}
\usage{
not_invisible_unit()
}
\description{
Test function that is explicitly NOT invisible.
}
