% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{raw_identifier_in_fn_args}
\alias{raw_identifier_in_fn_args}
\title{Test raw identifiers (\code{r#}) in function arguments are parsed correctly.
See (Issue #582)\link{https://github.com/extendr/extendr/issues/528} for details.}
\usage{
raw_identifier_in_fn_args(type = NULL)
}
\description{
Test raw identifiers (\code{r#}) in function arguments are parsed correctly.
See (Issue #582)\link{https://github.com/extendr/extendr/issues/528} for details.
}
