% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{special_param_names}
\alias{special_param_names}
\title{Test whether \verb{_arg} parameters are treated correctly in R}
\usage{
special_param_names(`_x`, `_y`)
}
\arguments{
\item{_x}{an integer scalar, ignored}

\item{_y}{an integer scalar, ignored}
}
\description{
Test whether \verb{_arg} parameters are treated correctly in R
}
