% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{true}
\alias{true}
\title{Test raw identifiers (\code{r#}) as function names are parsed correctly.
See (Issue #582)\link{https://github.com/extendr/extendr/issues/528} for details.}
\usage{
true()
}
\description{
Test raw identifiers (\code{r#}) as function names are parsed correctly.
See (Issue #582)\link{https://github.com/extendr/extendr/issues/528} for details.
}
