% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{try_char_vec}
\alias{try_char_vec}
\title{Convert a vector of strings to itself}
\usage{
try_char_vec(x)
}
\arguments{
\item{x}{a vector of strings}
}
\description{
Convert a vector of strings to itself
}
