% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{try_double_scalar}
\alias{try_double_scalar}
\title{Convert a double scalar to itself}
\usage{
try_double_scalar(x)
}
\arguments{
\item{x}{a number}
}
\description{
Convert a double scalar to itself
}
