% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{try_double_vec}
\alias{try_double_vec}
\title{Convert a vector of doubles to itself}
\usage{
try_double_vec(x)
}
\arguments{
\item{x}{a vector of doubles}
}
\description{
Convert a vector of doubles to itself
}
