% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{try_list_str_hash}
\alias{try_list_str_hash}
\title{Convert a list to a HashMap<&str, Robj> in rust and back. Does not preserve list order}
\usage{
try_list_str_hash(x)
}
\arguments{
\item{x}{a list}
}
\description{
Convert a list to a HashMap<&str, Robj> in rust and back. Does not preserve list order
}
