% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Wrapper}
\alias{Wrapper}
\alias{$.Wrapper}
\title{Class for testing (exported)}
\format{
An object of class \code{environment} of length 11.
}
\usage{
Wrapper
}
\description{
Class for testing (exported)
}
\examples{
x <- Wrapper$new()
x$a()
x$set_a(10)
x$a()
}
\keyword{datasets}
