% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hello_world}
\alias{hello_world}
\alias{get_extendrtests_metadata}
\alias{make_extendrtests_wrappers}
\title{Wrappers for Rust test functions.}
\usage{
hello_world()

get_extendrtests_metadata()

make_extendrtests_wrappers(use_symbols = TRUE)
}
\arguments{
\item{use_symbols}{use symbols instead of strings.}
}
\description{
Wrappers for Rust test functions.

Get module metadata as a list.

Build wrappers
}
