% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Regenerate extendr wrappers}
\usage{
document(pkg = ".")
}
\arguments{
\item{pkg}{Path to the package root. Defaults to current directory.}
}
\description{
Regenerates R wrappers from the Rust library by calling the
\verb{exported_make_\{package\}_wrappers} function exported by the wrappers shlib.
}
\details{
This function searches for the wrappers shlib in multiple locations:
\itemize{
\item Source builds: \verb{src/rust/target/\{debug,release\}/wrappers/}
\item Cross-compilation: \verb{src/rust/target/\{target-triple\}/\{debug,release\}/wrappers/}
\item Installed package: \verb{\{libpath\}/wrappers/}
}

When multiple candidates exist, the most recently modified one is used.

The wrappers shlib is a C wrapper around the Rust staticlib that exports
\verb{exported_make_\{package\}_wrappers}. This allows regenerating R wrappers without
requiring a cdylib crate type.
}
\keyword{internal}
