% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{MyClass}
\alias{MyClass}
\alias{$.MyClass}
\title{Class for testing (exported)}
\format{
An object of class \code{environment} of length 4.
}
\usage{
MyClass
}
\description{
Class for testing (exported)
}
\examples{
x <- MyClass$new()
x$get_a()
x$set_a(10)
x$get_a()
}
\keyword{datasets}
